backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.18
pydantic-settings>=2.3.0
alembic>=1.13.0
openai>=1.35.0
chromadb>=0.5.0
apscheduler>=3.10.0
fastmcp>=0.1.0
python-multipart>=0.0.9
httpx>=0.27.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1,<4.1.0
python-jose[cryptography]>=3.3.0
psycopg2-binary>=2.9.0
